paloalto/pyproject.toml
[project]
name = "paloalto"
version = "0.1.0"
description = "Palo Alto Panorama MCP Server for log analysis"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "azure-identity>=1.25.3",
    "azure-keyvault-secrets>=4.11.0",
    "azure-functions>=1.19.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.27.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "xmltodict>=1.0.4",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "services*", "tools*"]
exclude = ["__pycache__"]
